pyproject.toml
[project]
name = "diffulab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.4.0",
    "diffusers>=0.33.1",
    "einops>=0.8.1",
    "ema-pytorch>=0.7.7",
    "hydra-core>=1.3.2",
    "jaxtyping>=0.3.0",
    "mosaicml-streaming>=0.12.0",
    "omegaconf>=2.3.0",
    "open-clip-torch>=2.30.0",
    "pyopenssl==23.2.0",
    "torch>=2.6.0",
    "transformers>=4.49.0",
    "wandb>=0.19.6",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.1.1",
    "pre-commit>=4.2.0",
    "pyright>=1.1.396",
    "ruff>=0.9.7",
    "typos>=1.30.0",
]
repa = [
    "timm>=1.0.15",
]

[tool.uv.sources]
diffulab = {workspace = true}

[tool.astral]
package = "src"

[tool.black]
line-length = 120

[tool.ruff]
src = ["src", "examples"]
line-length = 120

[tool.ruff.lint]
select = [
    "I", # isort
]
ignore = [
    "F722", # forward-annotation-syntax-error, because of Jaxtyping
    "E731", # do-not-assign-lambda
]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.pyright]
include = ["src/diffulab", "tests/", "examples/"]
strict = ["*"]
exclude = ["**/__pycache__"]
reportMissingTypeStubs = "warning"
